nerin_final_updated/import_calc_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
sqlmodel==0.0.22
SQLAlchemy==2.0.36
pytz==2024.1
openpyxl==3.1.5
PyYAML==6.0.2
python-multipart==0.0.9
pytest==8.3.2
httpx==0.27.2
